% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{add_words_check}
\alias{add_words_check}
\title{check if vector or unit. If unit, break it into a vector}
\usage{
add_words_check(input)
}
\description{
check if vector or unit. If unit, break it into a vector
}
