% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{all_words}
\alias{all_words}
\title{returns a vector with words of a language.
The intent behind it is to test regex patterns}
\usage{
all_words(lang)
}
\description{
returns a vector with words of a language.
The intent behind it is to test regex patterns
}
