% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicio_pt.R
\name{build_dict}
\alias{build_dict}
\title{read a file .dic from Hunspell and get only the words to build a dict in the language}
\usage{
build_dict(arqs)
}
\arguments{
\item{arqs}{the name of one file
\dontrun{
grep("dic$",x = arqs, value = T)  |> build_dict()  |> head(5)
}}
}
\description{
read a file .dic from Hunspell and get only the words to build a dict in the language
}
