% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_words.R
\name{compound_words}
\alias{compound_words}
\title{Make compounded words into one word}
\usage{
compound_words(txt, subs)
}
\arguments{
\item{txt}{text to be processed}

\item{subs}{string with compounded words. E.g. "New_York Soviet_Union"}
}
\description{
Transform compounded words into one. E.g. "At New York", "It was in Soviet
Union" becomes "At New_York" and "It was in Soviet_Union". The compounded words
must be supplied in parameter `subs`. It can be used when counting words,
ploting wordclouds, cooccurence networks, help in naives Bayes
text classification.
}
\examples{
# text
txt <- "ice-cream in Soviet Union or New York?"
# words to be taken as a single
subs <- "Soviet_Union New_York"
compound_words(txt, subs)

# or using a vector of words
subs <- c("Soviet Union", "New York")
txt2 <- compound_words(txt, subs)
txt2
txt2 |> strsplit(" ")
}
