% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link.R
\name{connectors}
\alias{connectors}
\title{A lowercase connectors between two proper names}
\usage{
connectors(lang = "pt")
}
\arguments{
\item{lang}{language, It can be en, es, pt and misc (with many languages)}
}
\description{
In some languages there is a lowercase connector between two or more proper names.
This function returns a regex pattern by language with lowercase allowed connectors.
}
\examples{

connectors("es")
connectors("pt")
connectors("port")
connectors("en")
connectors("misc")
}
