% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{extract_abbrev}
\alias{extract_abbrev}
\title{Extract abbreviations from text}
\usage{
extract_abbrev(txt, stopwords = "", connectors = "do|de|o|a")
}
\arguments{
\item{txt}{input text}

\item{stopwords}{a vector of stop words. Pay attention to the capitalization.}
}
\description{
Extract abbreviations from text
}
\examples{
"Lorem Ipsum STF ergo S.T.F. foo." |> extract_abbrev()
"Lorem Ipsum FEDERAL SERVICE dolor sit S.T.F. foo." |> extract_abbrev()
"Lorem THE FEDERAL SERVICE dolor sit S.T.F. foo." |> extract_abbrev()
# now using the stopwords:
"Lorem THE FEDERAL SERVICE dolor sit S.T.F. foo." |> extract_abbrev(stopwords = "THE")
"Lorem THE FEDERAL SERVICE dolor sit WE S.T.F. foo." |> extract_abbrev(stopwords = c("THE", "WE"))
}
