% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link.R
\name{extract_graph}
\alias{extract_graph}
\title{Extract a non directional graph based on co-occurrence in the token.
It extracts only if two entities are mentioned in the same token (sentence or paragraph)}
\usage{
extract_graph(
  text,
  using = "sentences",
  connect = connectors("misc"),
  sw = gen_stopwords("en"),
  loop = FALSE
)
}
\arguments{
\item{text}{an input text}

\item{using}{sentence or paragraph to tokenize}

\item{connect}{lowercase connectors, like the "von" in "John von Neumann".}

\item{sw}{stopwords vector.}

\item{loop}{if TRUE, it will not remove loops, a node pointing to itself.}
}
\description{
Extract a non directional graph based on co-occurrence in the token.
It extracts only if two entities are mentioned in the same token (sentence or paragraph)
}
\examples{
text <- "John Does lives in New York in United States of America. He  is a passionate jazz musician, often playing in local clubs."
extract_graph(text)
}
