% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link.R
\name{extract_graph_rgx}
\alias{extract_graph_rgx}
\title{extract a graph from text, using custom regex pattern as nodes.}
\usage{
extract_graph_rgx(
  text,
  pattern,
  sw = gen_stopwords("en"),
  count_graphs = FALSE
)
}
\value{
a graph
}
\description{
extract a graph from text, using custom regex pattern as nodes.
}
