% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{extract_ppn}
\alias{extract_ppn}
\title{extract proper names from strings using regular expresssions.
The function suposes any sequence of upper letter followed by lower case is a proper name.
It is expected to return more things that wanted. Post-processing.}
\usage{
extract_ppn(string, connector = "d[aeo]s?")
}
\arguments{
\item{string}{input text}

\item{connector}{typical connector between names, like "of" "von", "van", "del". regex.}
}
\description{
extract proper names from strings using regular expresssions.
The function suposes any sequence of upper letter followed by lower case is a proper name.
It is expected to return more things that wanted. Post-processing.
}
\examples{
"O Joaquim José da Silva Xavier, tambén conhecido como Tiradentes foi um auferes." |> extract_ppn()
"José da Silva e Fulano de Tal foram, bla Maria Silva. E depois disso, bla Joaquim José da Silva Xavier no STF" |> extract_ppn()
}
