% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link.R
\name{extract_relation}
\alias{extract_relation}
\title{tokenize and selects only sentences/paragraphs with more than one entity per sentence or paragraph}
\usage{
extract_relation(
  text,
  using = "sentences",
  connectors = connectors("misc"),
  sw = gen_stopwords("en")
)
}
\description{
tokenize and selects only sentences/paragraphs with more than one entity per sentence or paragraph
}
\examples{
"John Does lives in New York in United States of America." |> extract_relation()
"João Ninguém mora em São José do Rio Preto. Ele foi para o Rio de Janeiro." |> extract_relation(connector = connectors("pt"))
}
