% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{gen_dict}
\alias{gen_dict}
\title{to generate a dictionary of specialized words
you can use regex and the function check the dictionary of the language and returns the matched words.
It is also useful to text your regex pattern.}
\usage{
gen_dict()
}
\description{
to generate a dictionary of specialized words
you can use regex and the function check the dictionary of the language and returns the matched words.
It is also useful to text your regex pattern.
}
