% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grep2}
\alias{grep2}
\title{A grep to be used with native pipe |>}
\usage{
grep2(x, arg1, ic = T, value = TRUE, ...)
}
\arguments{
\item{x}{data to search}

\item{arg1}{pattern}

\item{ic}{ignore case}

\item{...}{arguments passed to grep. See ?grep for more details}
}
\description{
grep2 is a wraper around the traditional grep, that works with native pipe,
returns value and are ignore case as default
}
\examples{
c("a", "b", "c", "d") |> grep2("a")
}
