% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grepl2}
\alias{grepl2}
\title{grepl to be used with native pipe |>}
\usage{
grepl2(x, arg1, ic = TRUE, ...)
}
\arguments{
\item{x}{data to search}

\item{arg1}{pattern}

\item{ic}{ignore case}

\item{...}{arguments passed to grepl.}
}
\description{
grepl to be used with native pipe |>
}
\examples{
c("a", "b", "c", "d") |> grepl2("a")
}
