% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{il}
\alias{il}
\title{Install libraries from string.}
\usage{
il(char, installer = "install.package", ...)
}
\arguments{
\item{char}{string with package names}

\item{installer}{choose which installer to be used: "install.package" (default), "pak", "dev_git" for devtools::install_github}

\item{...}{Additional parameters}
}
\description{
Given a char with package names, separated by spaces and/or commas, install all the packages
}
