% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ll}
\alias{ll}
\title{load libraries from string
load packages from a char separated by space or commas}
\usage{
ll(char, print = FALSE, ...)
}
\arguments{
\item{char}{character with package names}

\item{print}{if TRUE, will return a string that can be pasted in the console}

\item{...}{additional parameters}
}
\description{
load libraries from string
load packages from a char separated by space or commas
}
\examples{
sto::ll("rvest stringr dplyr")
sto::ll("rvest stringr dplyr", print=TRUE)
}
