% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ls2v}
\alias{ls2v}
\title{list of strings to list of vectors
string to vec from list of nested elements of text}
\usage{
ls2v(char, sep = " |\\\\n|\\\\t|\\\\r", wss = "_")
}
\description{
list of strings to list of vectors
string to vec from list of nested elements of text
}
