% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link.R
\name{plot_graph}
\alias{plot_graph}
\title{plot a network of coocurrence of terms}
\usage{
plot_graph(text, df, head_n = 30, color = "lightblue")
}
\arguments{
\item{text}{an input text}

\item{df}{a dataframe of co-occurrence, extracted with `extract_graph()` and `count(n1, n2)`}

\item{head_n}{number of nodes to show - the more frequent}
}
\description{
plot a graph of co-occurrence of terms, as returned by extract_graph
}
