% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_substitutions.R
\name{read_yml_dict}
\alias{read_yml_dict}
\title{Read YAML text as pure text and transform it into a named list}
\usage{
read_yml_dict(txt, sep = " ", ...)
}
\arguments{
\item{txt}{A character vector containing the YAML text to be read.}

\item{...}{additional parameters of s2v, like sep or keep_wss. see args(s2v)}
}
\value{
A named list where each element corresponds to a key-value pair in the YAML text.
}
\description{
To build your own complex dict of substitutions, create a text file with the
following yml like structure:
`key: word1 word2 word3` where word1 word2 word3 will be replaced by key in text.
The result of this function can be used with subs_synonyms() to replace
patterns in text.
}
\examples{
"Test: test teste TEST" |> read_yml_dict()
}
