% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{data}
\name{regex_NomeProprio}
\alias{regex_NomeProprio}
\title{a regex pattern to capture brazilian names, like "Fulano de Tal", "Ciclano dos Santos"}
\format{
An object of class \code{character} of length 1.
}
\usage{
regex_NomeProprio
}
\description{
a regex pattern to capture brazilian names, like "Fulano de Tal", "Ciclano dos Santos"
}
\keyword{datasets}
