% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexes.R
\docType{data}
\name{rgx_email}
\alias{rgx_email}
\title{regex to capture e-mail}
\format{
An object of class \code{character} of length 1.
}
\usage{
rgx_email
}
\description{
regex to capture e-mail
}
\examples{
text <- "Contact us at support@example.com or john.doe@company.co.no"
stringr::str_extract_all(text, rgx_email)[[1]]
}
\keyword{datasets}
