% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexes.R
\docType{data}
\name{rgx_url}
\alias{rgx_url}
\title{regex to capture websites address}
\format{
An object of class \code{character} of length 1.
}
\usage{
rgx_url
}
\description{
regex to capture websites address
}
\examples{
text <- "Contact us at www.example.com or https://www.john.doe.com"
stringr::str_extract_all(text, rgx_url)[[1]]
}
\keyword{datasets}
