% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2ppn.R
\name{s2ppn}
\alias{s2ppn}
\title{Transform the text into proper name style}
\usage{
s2ppn(text, prep = s2v("da de di do das dos von van of the"))
}
\arguments{
\item{text}{the text to be transformed into name style}

\item{prep}{a vector of prepositions to be in lower case}
}
\description{
Traditional funcions like toupper() or tolower() transform the text into
upper case or lower case. The stringr::str_to_title() transforms all text into lower case, with the exception of the first word. This function transform the text into proper name style.
}
\examples{
"augusto dos anjos" |> s2ppn(prep = "dos")
"AUGUSTO DOS ANJOS" |> s2ppn(prep = "dos")
"AUGUSTO Dos anjos" |> s2ppn(prep = s2v("da de di do das dos"))
}
