% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{s2v}
\alias{s2v}
\title{Convert a string into a vector of elements.}
\usage{
s2v(
  char,
  sep = " |\\\\n|\\\\t|\\\\r|,|;",
  wss = "_",
  keep_wss = FALSE,
  print = FALSE
)
}
\arguments{
\item{char}{character to be transformed}

\item{sep}{separator. It can be a regular expression pattern.}

\item{wss}{whitespace substitution. Which character will be used in 
the end to be converted into white space}

\item{keep_wss}{keep separator. If TRUE, will keep the separator, 
if FALSE (default), it will be replaced by white space.}

\item{print}{if TRUE, will return a string that can be pasted in the console}
}
\description{
Easily transform a character string into a vector of elements.
}
\examples{
"a b c d" |> s2v(r
# strips white spaces
"a b c     d" |> s2v()
"a b c\nd\te" |> s2v()
# And specifying the separator character:
"a|b|c|d e" |> s2v(sep = "\\\\|")
# To use the output verbatim as a string to copy and use in the code
"a b c d" |> s2v(print = TRUE)
}
