% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{s_extract_all}
\alias{s_extract_all}
\title{extract all chars.}
\usage{
s_extract_all(txt, pattern, IC = TRUE, unl = FALSE)
}
\description{
Inspired by stringr::str_extract_all(), but with ignore case option.
}
\examples{
"bla bla bla Foo ble ble" |> s_extract_all("foo")
c("bla bla bla Foo ble", "Lorem FOO ipsum") |> s_extract_all("foo", unl = T)
}
