% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shuffle_time}
\alias{shuffle_time}
\title{generate shuffle times
easily generate shuffle time, good to use in webscraping}
\usage{
shuffle_time(min_time, max_time)
}
\description{
generate shuffle times
easily generate shuffle time, good to use in webscraping
}
