% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strip_txt}
\alias{strip_txt}
\title{funcao para apagar o vetor de texto a partir do início de certo padrão de txt}
\usage{
strip_txt(txt, criteria, to_end = TRUE)
}
\arguments{
\item{to_end}{If TRUE, erases from criteria to the end of text. if FALSE, erases from the begin to the criteria}
}
\description{
funcao para apagar o vetor de texto a partir do início de certo padrão de txt
}
