% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_substitutions.R
\name{subs_synonyms}
\alias{subs_synonyms}
\title{Replace different words by a single word}
\usage{
subs_synonyms(txt, subst, ic = FALSE)
}
\arguments{
\item{txt}{The input text where replacements need to be made.}

\item{subst}{the substitutions. It accepts as input 1) a dataframe, where
the column one is the pattern to capture the words do be replaced and the
second column is the word that will replace the first; 2) A named list where
keys are the target words and values are vectors of synonyms. Normally, this
list is generated by function read_yml_dict()}

\item{ic}{ignore case? default FALSE
it works great with function read_yml_dict(), that creates a list from yaml file.}
}
\value{
a list with key that will substitute the elements of list element.
}
\description{
When dealing with text analysis, a thing may have different strings, or synonyms.
This function make it easier to replace in text many synonym words to only one.
}
\examples{
:
dic_subs <- "WEF: World_Economic_Forum W\\\\.E\\\\.F\\\\."
subst <- read_yml_dict(dic_subs)
txt <- "The World Economic Forum announced that X and Y. The W.E.F. is a lorem ipsum. But the world economic forum lorem ipsum."
subs_synonyms(txt, subst)
subs_synonyms(txt, subst, ic = TRUE)
}
