% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_words.R
\name{compound_words}
\alias{compound_words}
\title{Make compounded words into one word}
\usage{
compound_words(txt, subs)
}
\arguments{
\item{txt}{text to be processed}

\item{subs}{string with compounded words. E.g. "New_York Soviet_Union"}
}
\description{
Transform compounded words into one. E.g. "At New York", "It was in Soviet Union" becomes "At New_York" and "It was in Soviet_Union". The compunded words must be supplied in parameter `subs`.
}
\examples{
txt <- "ice-cream in Soviet Union or New York?"
s <- "Soviet_Union New_York"
compounded_words(txt, s)
}
