% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f}
\alias{f}
\title{To easily paste and collpase chars objects into one string
A wrapper for glue::glue}
\usage{
f(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .null = character(),
  .comment = "#",
  .literal = FALSE,
  .transformer = identity_transformer,
  .trim = TRUE
)
}
\description{
To easily paste and collpase chars objects into one string
A wrapper for glue::glue
}
