% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ll}
\alias{ll}
\title{load libraries from string
load packages from a char separated by space or commas}
\usage{
ll(char, ...)
}
\description{
load libraries from string
load packages from a char separated by space or commas
}
\examples{
sto::ll("rvest stringr dplyr")
}
