% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nothing}
\alias{nothing}
\title{An empty function.}
\usage{
nothing(input)
}
\description{
An empty function that does nothing, just returns the input. It is useful when developing and testing pipes, placed in the end of pipes, so the developer can comment and uncomment lines of pipes easily, instead of adding or deleting the last coma or pipe in the last line.
}
\examples{
s2v("a b c d 1 2 3 4") |>
grep2("[[:alpha:]]") |> # select all letters
# grep2("[[:digit:]]")  |> # select all numbers
nothing()

# Changing the commented lines: 
s2v("a b c d 1 2 3 4") |>
# grep2("[[:alpha:]]") |> # select all letters
grep2("[[:digit:]]")  |> # select all numbers
nothing()
}
