% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2df.R
\name{s2df}
\alias{s2df}
\title{Convert a string in CSV like format into a dataframe/tibble}
\usage{
s2df(input_string, header = TRUE, sep = ";")
}
\arguments{
\item{input_string}{A character string in the format "col1;col2\nvalue1;value2".
The first line should be the header row.}
}
\value{
A `tibble` with columns named according to the header and rows populated with data from the subsequent lines.
}
\description{
Convert a string in CSV like format into a dataframe/tibble
}
\examples{
#'
s <- "col1;col2
bla1;10
bla2;23
bla3;32
"
s2df(s)

s2 <- "bla1,10\nbla2,23\nbla3,32"
s2df(s2, header = FALSE, sep = ",")
}
