% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{show_sw}
\alias{show_sw}
\title{show all stopwords categories of a language}
\usage{
show_sw(lang, as_vector = FALSE)
}
\arguments{
\item{lang}{language, like "en", "pt"}
}
\value{
an list object of stopwords
}
\description{
show all stopwords categories of a language
}
\examples{
show_sw("pt")
show_sw("en")
}
